perf: change transactions type to `Vec<Vec<u32>>`
`usize` takes 64 bit on a 64 bit target (which is pervasive).
Changing to u32 can reduce some memory usage, from what I've
expiremented, 1.84GiB memory allocations reduced to 1.82GiB.

+++ bench/src/main.rs
@@ -30,7 +30,7 @@ fn main() {
     for line in reader.lines() {
         if let Ok(line) = line {
             let line = line.split_whitespace().collect::<Vec<&str>>();
-            transactions.push(line.iter().map(|x| x.parse::<usize>().unwrap()).collect());
+            transactions.push(line.iter().map(|x| x.parse::<u32>().unwrap()).collect());
         }
     }
